+++ api/mon/backends/dummy/__init__.py
@@ -1,9 +1,17 @@
 from api.mon.backends.abstract import AbstractMonitoringBackend
+from api.mon.backends.abstract.server import AbstractMonitoringServer
 
 
 # noinspection PyAbstractClass
 class DummyMonitoring(AbstractMonitoringBackend):
     pass
+
+
+class DummyMonitoringServer(AbstractMonitoringServer):
+    """
+    Dummy model for representing a monitoring server in a DC.
+    """
+    Meta = AbstractMonitoringServer.Meta
 
 
 def get_monitoring(dc, **kwargs):
@@ -16,3 +24,4 @@ def del_monitoring(dc):
 
 
 MonitoringBackendClass = DummyMonitoring
+MonitoringServerClass = DummyMonitoringServer